# server/swagger/swagger.yaml
basePath: /v1
definitions:
  bodies.ClientLogCreationRequest:
    properties:
      key:
        type: string
      platform:
        maxLength: 64
        type: string
      type:
        description: Needs manual validation for enum
        maxLength: 32
        type: string
    required:
      - key
      - platform
      - type
    type: object
  bodies.ClientLogCreationResponse:
    properties:
      expiry:
        type: string
      link_web:
        type: string
      log_id:
        type: string
      token:
        type: string
    type: object
  bodies.ErrorResponse:
    properties:
      details:
        description: |-
          Details contains details about the error.
          This field is mostly for development, but is used in production for adding request body validation errors.
        type: string
      error:
        description: Error contains the error string.
        type: string
      raw_error:
        description: |-
          RawError contains the raw error.
          This field is only sent in development mode.
        type: string
    type: object
  bodies.LogMetadataSaveRequest:
    properties:
      additional_metadata:
        additionalProperties: { }
        type: object
      log_links:
        additionalProperties:
          type: string
        type: object
    type: object
  bodies.StatusResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: ok
        type: string
      version:
        example: 0.0.1
        type: string
    type: object
  bodies.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 64
        minLength: 12
        type: string
    required:
      - email
      - password
    type: object
  bodies.UserPasswordResetCompleteRequest:
    properties:
      password:
        maxLength: 64
        minLength: 12
        type: string
      resetToken:
        type: string
    required:
      - password
      - resetToken
    type: object
  entities.LogAdditionalMetadata:
    properties:
      metadata:
        additionalProperties: { }
        type: object
      savedOn:
        type: string
    type: object
  entities.LogLink:
    properties:
      description:
        type: string
      linkedLog:
        type: string
    type: object
  entities.LogMetadata:
    properties:
      additionalMetadata:
        items:
          $ref: '#/definitions/entities.LogAdditionalMetadata'
        type: array
      linkedLogs:
        items:
          $ref: '#/definitions/entities.LogLink'
        type: array
    type: object
  entities.LogOverview:
    properties:
      id:
        type: string
      lineCount:
        type: integer
      logType:
        type: string
      platform:
        type: string
      tsFirstLine:
        type: string
      tsLastLine:
        type: string
    type: object
  entities.TitleEnvironment:
    properties:
      environmentKey:
        type: string
      environmentName:
        type: string
      title:
        type: string
    type: object
  entities.UserLoginResult:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      token:
        type: string
      tokenExpiry:
        type: string
      userUUID:
        type: string
    type: object
host: example.com
info:
  contact:
    email: no-reply@capsa.gg
    name: Luciano Nooijen
    url: https://lucianonooijen.com
  description: Storage server for Project Capsa, a ground-breaking Unreal Engine logging
    solution
  title: Capsa Server
  version: "1"
paths:
  /client/auth:
    post:
      consumes:
        - application/json
      description: Allows clients to create a log session and receive a token to send
        logs with
      parameters:
        - description: ClientLogCreationRequest
          in: body
          name: creation_request
          required: true
          schema:
            $ref: '#/definitions/bodies.ClientLogCreationRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ClientLogCreationResponse'
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "401":
          description: Unauthorized
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "403":
          description: Forbidden
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "409":
          description: Conflict
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      summary: Client authentication handler
      tags:
        - Client
  /client/log/chunk:
    post:
      consumes:
        - text/plain
      description: Allows clients to upload log chunks for their log sessions. To
        test this endpoint, please run the upload locally to add the correct request
        body.
      parameters:
        - description: Plain text log
          in: body
          name: log
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
        "304":
          description: Not Modified
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "415":
          description: Unsupported Media Type
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      security:
        - JwtClient: [ ]
      summary: Log chunk storage
      tags:
        - ClientAuthenticated
  /client/log/metadata:
    post:
      consumes:
        - application/json
      description: Allows clients to store linked logs and additional metadata for
        their log session
      parameters:
        - description: LogMetadataSaveRequest
          in: body
          name: metadata
          required: true
          schema:
            $ref: '#/definitions/bodies.LogMetadataSaveRequest'
      responses:
        "201":
          description: Created
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      security:
        - JwtClient: [ ]
      summary: Log metadata storage
      tags:
        - ClientAuthenticated
  /status:
    get:
      description: To be used with for status pings to check readiness and liveliness
        of the server
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.StatusResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.StatusResponse'
      summary: Status handler
      tags:
        - Common
  /user/auth/login:
    post:
      consumes:
        - application/json
      description: Allows clients to create a log session and receive a token to send
        logs with
      parameters:
        - description: UserLoginRequest
          in: body
          name: login_request
          required: true
          schema:
            $ref: '#/definitions/bodies.UserLoginRequest'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/entities.UserLoginResult'
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "401":
          description: Unauthorized
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "403":
          description: Forbidden
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "409":
          description: Conflict
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      summary: User login handler
      tags:
        - UserAuthentication
  /user/auth/password-reset:
    get:
      description: Allows clients to create a log session and receive a token to send
        logs with
      parameters:
        - description: UserEmail
          in: query
          name: email
          required: true
          type: string
      produces:
        - application/json
      responses:
        "201":
          description: Created
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      summary: User password reset request
      tags:
        - UserAuthentication
    post:
      description: Allows clients to create a log session and receive a token to send
        logs with
      parameters:
        - description: UserEmail
          in: query
          name: email
          required: true
          type: string
        - description: UserPasswordCompleteRequest
          in: body
          name: password_reset_data
          required: true
          schema:
            $ref: '#/definitions/bodies.UserPasswordResetCompleteRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      summary: User password reset set password
      tags:
        - UserAuthentication
  /user/environments:
    get:
      description: Allows users to fetch all available environments from the database
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            items:
              $ref: '#/definitions/entities.TitleEnvironment'
            type: array
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      security:
        - JwtUser: [ ]
      summary: Environment listing
      tags:
        - UserEnvironments
  /user/logs:
    get:
      description: Allows users to fetch available logs from the database
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            items:
              $ref: '#/definitions/entities.LogOverview'
            type: array
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      security:
        - JwtUser: [ ]
      summary: Log listing
      tags:
        - UserLogs
  /user/logs/{logid}/log:
    get:
      description: Allows users to stream all uploaded chunks for a given log id
      produces:
        - text/plain
        - application/json
      responses:
        "200":
          description: Log chunk stream
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            type: string
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Error:
              description: Server error information
              type: string
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      security:
        - JwtUser: [ ]
      summary: Log chunk storage
      tags:
        - UserLogs
  /user/logs/{logid}/metadata:
    get:
      description: Allows users to fetch metadata for a log
      produces:
        - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/entities.LogMetadata'
        "400":
          description: Bad Request
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "404":
          description: Not Found
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
        "500":
          description: Internal Server Error
          headers:
            X-Capsa-Server-Version:
              description: Current Capsa Server version
              type: string
          schema:
            $ref: '#/definitions/bodies.ErrorResponse'
      security:
        - JwtUser: [ ]
      summary: Log metadata
      tags:
        - UserLogs
schemes:
  - http
  - https
securityDefinitions:
  JwtClient:
    description: Header value should be "Bearer ClientJwtString"
    in: header
    name: Authorization
    type: apiKey
  JwtUser:
    description: Header value should be "Bearer UserJwtString"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"